pyproject.toml
[project]
name = "webshare-proxy-generator"
version = "0.1.0"
description = "Gens proxys from webshare for free 10 at a time async"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "httpx>=0.28.1",
    "nicegui>=2.21.1",
    "numpy==2.2",
    "openai-whisper>=20250625",
    "playwright>=1.53.0",
    "playwright-stealth>=2.0.0",
    "psutil>=7.0.0",
    "pywin32>=310",
    "scipy>=1.16.0",
]
